Move module kas-core::event::cx::config -> kas-core::event::config_cx
